Updated to v2.3.0-stable.20230602
- Fixed some bugs related to recording listening data.
- Refactored `App.tsx` code to smaller modules.
-  Fixed some brightness issues in SongCard.
