+++ src/renderer/components/ArtistInfoPage/DuplicateArtistsSuggestion.tsx
@@ -109,6 +109,7 @@ const DuplicateArtistsSuggestion = (props: Props) => {
           return addNewNotifications([
             {
               content: 'Artist conflict resolved successfully.',
+              iconName: 'done',
               delay: 5000,
               id: 'ArtistDuplicateSuggestion',
             },
